Changed project to use .meta files for version control.
Please update unity whenever possible, we will try to use the most recent version of unity whenever possible. If not possible, follow the steps below to open the project in an old version.

If using an old version of Unity (on a school computer, say), delete the Library, Temp folders and any Assembly and .sln files. Then reopen the project in unity and let it rebuild.